SequGen

Documentation for SequGen.

SequGen.Constant

SequGen.GaussianNoise

SequGen.Sequence

SequGen.SineWave

SequGen.StepwiseGaussianNoise

SequGen.Trend

SequGen.sample

SequGen.Constant — Type

Constant(x)

Constant sequence. Always returns x.

SequGen.GaussianNoise — Type

GaussianNoise(σ = 1.0)

Returns a Gaussian noise with standard deviation σ.

SequGen.Sequence — Type

Sequence

Base type for synthetic sequences.

SequGen.SineWave — Type

SineWave(amplitude, period, phase)
SineWave(:from_phase, amplitude, period, phase)
SineWave(:from_peak, amplitude, period, peak_time)

Creates a sine wave, i.e., something of the form

ϕ(t) = A sin( 2π t / T + θ₀),

where A is the amplitude, T is the period, and θ₀ is the phase.

SequGen.StepwiseGaussianNoise — Type

StepwiseGaussianNoise(stepsize; σ = 1.0)

Returns a Gaussian noise with standard deviation σ and the given stepsize. The noise is constant between interval of size stepsize. For instance,

sample(StepwiseGaussianNoise(1.0), 0.0, 2.0, 10)

will return a sample of 10 elements with 5 equal values followed by other 5 equal values. This is equivalent to [randn() * ones(5); randn() * ones(5)] in this case.

SequGen.Trend — Type

Trend(t₀, slope)

Constant slope starting at start, i.e., ϕ(t) = (t - t₀) * slope.

SequGen.sample — Method

x = sample(sequence, range)
x = sample(sequence, start, end, length)

Compute a sample in the given range, or in the range given by range(start, end, length=length).